pyproject.toml
[tool.poetry]
name = "weird-o"
version = "0.1.0"
description = ""
authors = ["tarannum-2002 <tarannums@sahaj.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
qrcode = {extras = ["pil"], version = "^8.0"}
opencv-python = "^4.10.0.84"
pillow = "^11.0.0"
pyzbar = "^0.1.9"
numpy = "^2.1.3"
matplotlib = "^3.9.3"
squiggle = "^0.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
